cplsa module deleted - classes moved to diversity and mf modules; comments to public classes and members added
