+++ RankSys-diversity/pom.xml
@@ -43,6 +43,11 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
             <scope>test</scope>
         </dependency>
         <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>RankSys-mf</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-math3</artifactId>
             <version>3.5</version>
