+++ RankSys-diversity/src/main/java/org/ranksys/diversity/intentaware/package-info.java
@@ -1,0 +1,8 @@
+/*
+ * Copyright (C) 2016 RankSys http://ranksys.org
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.ranksys.diversity.intentaware;
