+++ RankSys-examples/src/main/java/org/ranksys/examples/CPLSARerankerExample.java
@@ -1,3 +1,10 @@
+/*
+ * Copyright (C) 2016 RankSys http://ranksys.org
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
 package org.ranksys.examples;
 
 import es.uam.eps.ir.ranksys.diversity.intentaware.reranking.XQuAD;
@@ -25,6 +32,11 @@ import org.ranksys.formats.rec.SimpleRecommendationFormat;
 import static org.ranksys.formats.parsing.Parsers.lp;
 import static org.ranksys.formats.parsing.Parsers.sp;
 
+/**
+ * Example of CPLSA factorization usage as a source of models for the intent-aware diversification.
+ *
+ * @author Jacek Wasilewski (jacek.wasilewski@insight-centre.org)
+ */
 public class CPLSARerankerExample {
 
     public static void main(String[] args) throws Exception {
