+++ pom.xml
@@ -37,7 +37,6 @@ file, You can obtain one at http://mozilla.org/MPL/2.0/.
         <module>RankSys-compression</module>
         <module>RankSys-formats</module>
         <module>RankSys-examples</module>
-        <module>RankSys-cplsa</module>
     </modules>
     
     <build>
